(test) on failure, free variables in reverse order
Implement a queue in the optkit.tests.C.base module to preserve
order of C variables as they are registered with the TestCase
object during testing; if the test fails, registered variables
are freed in the reverse order of their addition. This way,
later variables that are dependent on memory allocated in
earlier ones cannot be freed first. The variables' ree()
methods themselves should not depend on earlier variables,
so this change is probably not necessary but seems sane.