(test) correct generation of input to equilibrate
1. For the regularized_sinkhorn_knopp() call used to equilibrate
dense matrices, the Python test passes in an argument of type

    ok_float *

which is intended to represent the input array. Generating this
pointer from a <numpy.array>, however, differs between OSX and
linux. On OSX, the chained command

    ptr = array.astype(<float_type>).ctypes.data_as(<float_ptr_type>)

yields a working pointer. On linux, this call somehow mangles
the pointer (like, introduces offset backwards or forwards or
something, such that the first few entries pointed to are
uninitialized gibberish). Instead, the two calls:

    good_array = array.astype(<float_type>)
    good_array_ptr = good_array.ctypes.data_as(<float_ptr_type>)

yield the desired result. Consequently, the cpu 32/64 equilibration
libraries that were passing on OSX and failing on linux now also
pass on linux.

2. Correct typo in optkit_equilibration.c: missing right parenthesis
after call to vector_uniform_rand()

3. Clean up error checking calls in optkit_equilibration.c

newly passing:
- OPERATOR (gpu 32/64)
- EQUIL (gpu 32/64, linux cpu 32/64)

failures:
- CG (linux cpu 32)

untested:
- CLUSTER (linux cpu 32/64, gpu 32/64)
- CG (gpu 32/64)
- PROJECTOR (gpu 32/64)
- POGS (gpu 32/64)
- ABSTRACT POGS (gpu 32/64)