(test) correct blas_handle creation in projector test
1. In test_projector.py unit tests, the BLAS handle variable was
being created once but free multiple times. This is fine on CPU where
BLAS handle creation and destruction are no-ops, but invalid on GPU.
Move handle creation inside the test loop (destruction at end of loop)
to resolve this problem. Compile, test & pass PROJECTOR on gpu 32/64.

2. Compile, test, pass CG on gpu 32/64.

3. libcg_cpu32 now passes on linux, was failing previously.
Not sure which upstream change this is attributable to.

newly passing:
- CG (gpu 32/64, linux cpu 32)
- PROJECTOR (gpu 32/64)

failures:
- { NONE }

untested:
- CLUSTER (linux cpu 32/64, gpu 32/64)
- POGS (gpu 32/64)
- ABSTRACT POGS (gpu 32/64)