(test) loosen pogs convergence tolerances for GPU
CPU test includes a 10x factor for primal residuals, 20x
factor for dual residuals to account for errors introduced
by unscaling the problem after the solver converges.

GPU test was failing under the same constraint; adding a
1.5x factor resolved this.

In other words, we expect

	||Ax - y|| <= atol * sqrt(m) + 10 * rtol * ||y||
	||A'nu + mu|| <= atol * sqrt(n) + 20 * rtol * ||mu||

to hold on the CPU, while

	||Ax - y|| <= atol * sqrt(m) + 15 * rtol * ||y||
	||A'nu + mu|| <= atol * sqrt(n) + 30 * rtol * ||mu||

to hold on the GPU, where atol and rtol are the absolute and
relative tolerances specified to the solver, m x n are the
dimensions of matrix A, and (x, y) and (mu, nu) are the
unscaled optimal primal and dual variables, respectively.

newly passing:
- POGS (gpu 32/64)
- ABSTRACT POGS (gpu 32/64)

failures:
- { NONE }

untested:
- CLUSTER (linux cpu 32/64, gpu 32/64)