(test) modify C test script exit code
TravisCI looks at exit code from script "./test";
if script runs nosetests, the success/failures of
the tests are not passed through (duh). so we
now gather this information and feed it through:
exit code set to 1 if any nosetests tests fail.

exemptions:
-test_cluster.py
-test_abstract_pogs.py
(exempt tests are run and logged, but failures ignored
for TravisCI)