(test) allow repeated projection tests
Testing with python 3.3 appears to miss
numerical tolerance for projection test results;
the check is to see whether
\[
	(x_proj, y_proj) \in \{(x,y) | y=Ax \}
\]
holds to a given tolerance. While $x_proj$ and
$y_proj$ are generated by C, they may be rounded
by Python, and the product $Ax_proj$ is formed
in Python and may be carried out to a different
tolerance in Py2.7 vs Py 3.3 (i.e. 32-bit vs.
64-bit float operations by default?)

Since $A$ is generated randomly, if the random
number generation differ across the Python
versions, the random $A$s obtained in Py3.3
may, for instance, have worse condition number,
impacting the output of the projection.

In any case, we allow for the condition
\[
	||Ax - y|| <=
	\epsilon_abs + \epsilon_rel * ||y||
\]
to hold less strictly (multiplying the tolerances
by a factor of 2 on the second and third repeated
trials) or to hold for different matrices
(up to 5 total attempts).