(vector/matrix) propage array viewing safeguards
in previous commit, the functions
	vector_subvector
	matrix_submatrix
	matrix_column
	matrix_row
	matrix_diagonal
were modified to add a safeguard: if the (vector or matrix)
argument provided to serve as the container of the array view
already has allocated data attached, an error is returned.

the purpose of this is to avoid overwriting a pointer to
allocated data. however, in cases in which the vector
is already a viewer of data that it does not own, the
error will still be generated since the .data field
is not null. therefore, the client must manage the .data
field by resetting it to null when the view is no longer
needed, and prior to invoking a new viewing call.

this commit propagates the correct management of
array views to the clients of the above functions.